refactor(notifications): remove leftover opt-in backward-compat
Drop unused prompt hooks and route remaining after-action prompts through notifyFlowCompleted.

LIVE-25271
