+++ .changeset/quiet-drawers-retire.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": patch
+---
+
+Remove leftover notifications opt-in backward-compat and unused prompt hooks
