+++ apps/ledger-live-mobile/src/mvvm/features/AssetDetail/screens/AssetDetail/components/CoinOptions/__tests__/useAssetCoinOptionsViewModel.test.ts
@@ -2,11 +2,8 @@ import { act, renderHook } from "@tests/test-renderer";
 import { getCryptoCurrencyById } from "@domain/entity-currency-crypto";
 import { track } from "~/analytics";
 import type { State } from "~/reducers/types";
+import { NotificationsPromptProvider } from "LLM/features/NotificationsPrompt";
 import { useAssetCoinOptionsViewModel } from "../useAssetCoinOptionsViewModel";
-
-jest.mock("LLM/features/NotificationsPrompt", () => ({
-  useNotifications: () => ({ tryTriggerPushNotificationDrawerAfterAction: jest.fn() }),
-}));
 
 jest.mock("~/analytics", () => ({ track: jest.fn() }));
 
@@ -26,6 +23,7 @@ function renderViewModel({
   return renderHook(
     () => useAssetCoinOptionsViewModel({ currency: bitcoin, currencyId, marketId }),
     {
+      innerWrapper: NotificationsPromptProvider,
       overrideInitialState: (state: State): State => ({
         ...state,
         settings: { ...state.settings, blacklistedTokenIds, starredMarketCoins },
@@ -44,7 +42,7 @@ describe("useAssetCoinOptionsViewModel", () => {
 
     act(() => result.current.onToggleFavourite());
     expect(store.getState().settings.starredMarketCoins).toContain(bitcoin.id);
-    expect(track).toHaveBeenLastCalledWith(
+    expect(track).toHaveBeenCalledWith(
       "button_clicked",
       expect.objectContaining({
         button: "favourite",
@@ -54,7 +52,7 @@ describe("useAssetCoinOptionsViewModel", () => {
 
     act(() => result.current.onToggleFavourite());
     expect(store.getState().settings.starredMarketCoins).not.toContain(bitcoin.id);
-    expect(track).toHaveBeenLastCalledWith(
+    expect(track).toHaveBeenCalledWith(
       "button_clicked",
       expect.objectContaining({
         button: "favourite",
@@ -72,6 +70,7 @@ describe("useAssetCoinOptionsViewModel", () => {
           marketId: "shiba-inu",
         }),
       {
+        innerWrapper: NotificationsPromptProvider,
         overrideInitialState: (state: State): State => ({
           ...state,
           settings: { ...state.settings, starredMarketCoins: [] },
