+++ apps/ledger-live-mobile/src/mvvm/features/NotificationsPrompt/new/NotificationsPromptProvider.tsx
@@ -1,17 +1,17 @@
-import React, { createContext, useContext, useMemo } from "react";
-import {
-  type InitPushNotificationsDataResult,
-  type NotificationsPromptAfterActionSource,
-} from "LLM/features/NotificationsPrompt";
+import React, { createContext, useContext, useEffect, useMemo } from "react";
+import { AppState } from "react-native";
+import type { InitPushNotificationsDataResult } from "LLM/features/NotificationsPrompt/types";
+import type { NotificationsPromptAfterActionSource } from "LLM/features/NotificationsPrompt/utils/notificationsPromptEngine";
 import { useNotificationsPromptTriggers } from "LLM/features/NotificationsPrompt/new/hooks/useNotificationsPromptTriggers";
 
 type NotificationsPromptProviderProps = {
-  children: React.ReactNode;
+  children?: React.ReactNode;
 };
 
 export type NotificationsPromptContextValue = {
   notifyFlowCompleted: (source: NotificationsPromptAfterActionSource) => void;
   tryTriggerPushNotificationDrawerAfterInactivity: (data: InitPushNotificationsDataResult) => void;
+  initPushNotificationsData: () => Promise<InitPushNotificationsDataResult>;
 };
 
 export const NotificationsPromptContext = createContext<NotificationsPromptContextValue | null>(
@@ -29,15 +29,39 @@ export function useNotificationsPrompt() {
 }
 
 export function NotificationsPromptProvider({ children }: NotificationsPromptProviderProps) {
-  const { notifyFlowCompleted, tryTriggerPushNotificationDrawerAfterInactivity } =
-    useNotificationsPromptTriggers();
+  const {
+    notifyFlowCompleted,
+    tryTriggerPushNotificationDrawerAfterInactivity,
+    initPushNotificationsData,
+  } = useNotificationsPromptTriggers();
+
+  useEffect(() => {
+    initPushNotificationsData().then(tryTriggerPushNotificationDrawerAfterInactivity);
+    // oxlint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
+  // Re-read OS permission and stored opt-in data when the app returns to the foreground.
+  useEffect(() => {
+    const subscription = AppState.addEventListener("change", nextAppState => {
+      if (nextAppState === "active") {
+        initPushNotificationsData();
+      }
+    });
+
+    return () => subscription.remove();
+  }, [initPushNotificationsData]);
 
   const value = useMemo(
     () => ({
       notifyFlowCompleted,
       tryTriggerPushNotificationDrawerAfterInactivity,
+      initPushNotificationsData,
     }),
-    [notifyFlowCompleted, tryTriggerPushNotificationDrawerAfterInactivity],
+    [
+      notifyFlowCompleted,
+      tryTriggerPushNotificationDrawerAfterInactivity,
+      initPushNotificationsData,
+    ],
   );
 
   return (
