+++ apps/ledger-live-mobile/src/mvvm/features/NotificationsPrompt/new/hooks/useNotificationsPromptTriggers.ts
@@ -3,10 +3,12 @@ import { useFeature } from "@features/platform-feature-flags";
 import { useSelector } from "~/context/hooks";
 import { ratingsModalOpenSelector } from "~/reducers/ratings";
 import { hasCompletedOnboardingSelector, notificationsSelector } from "~/reducers/settings";
+import { getNotificationPermissionStatus } from "~/logic/getNotificationPermissionStatus";
 import { useNotificationsPermission } from "LLM/hooks/useNotificationsPermission";
 import {
   evaluateAfterActionTrigger,
   evaluateInactivityTrigger,
+  getPushNotificationsDataOfUserFromStorage,
   type InitPushNotificationsDataResult,
   type NotificationsPromptAfterActionSource,
   useNotificationsData,
@@ -22,9 +24,83 @@ export function useNotificationsPromptTriggers() {
   const hasCompletedOnboarding = useSelector(hasCompletedOnboardingSelector);
   const notifications = useSelector(notificationsSelector);
   const isRatingsModalOpen = useSelector(ratingsModalOpenSelector);
-  const { permissionStatus } = useNotificationsPermission();
-  const { pushNotificationsDataOfUser } = useNotificationsData();
+  const { permissionStatus, setPermissionStatus } = useNotificationsPermission();
+  const {
+    pushNotificationsDataOfUser,
+    initializeNotificationSettingsState,
+    syncOptOutState,
+    updatePushNotificationsDataOfUserInStateAndStore,
+  } = useNotificationsData();
   const { openDrawer, isDrawerPending } = useNotificationsPromptDrawerScheduler();
+
+  const initPushNotificationsData =
+    useCallback(async (): Promise<InitPushNotificationsDataResult> => {
+      initializeNotificationSettingsState();
+
+      const [permission, dataOfUserFromStorage] = await Promise.allSettled([
+        getNotificationPermissionStatus(),
+        getPushNotificationsDataOfUserFromStorage(),
+      ]);
+
+      if (permission.status === "rejected") {
+        console.error("Failed to get notification permission status:", permission.reason);
+      }
+
+      if (dataOfUserFromStorage.status === "rejected") {
+        console.error(
+          "Failed to get push notifications user data from storage:",
+          dataOfUserFromStorage.reason,
+        );
+      }
+
+      if (dataOfUserFromStorage.status === "fulfilled") {
+        const storedUserData = dataOfUserFromStorage.value;
+
+        if (permission.status === "fulfilled") {
+          const osPermissionStatus = permission.value;
+
+          setPermissionStatus(osPermissionStatus);
+
+          syncOptOutState(osPermissionStatus, storedUserData);
+          return {
+            status: "success",
+            storedUserData,
+            osPermissionStatus,
+            areAppNotificationsEnabled: notifications.areNotificationsAllowed,
+          };
+        }
+
+        if (permission.status === "rejected") {
+          updatePushNotificationsDataOfUserInStateAndStore(storedUserData ?? {});
+          return {
+            status: "error",
+            reason: "Failed to get notification permission status",
+          };
+        }
+      }
+
+      if (dataOfUserFromStorage.status === "rejected" && permission.status === "fulfilled") {
+        const osPermissionStatus = permission.value;
+        setPermissionStatus(osPermissionStatus);
+
+        return {
+          status: "error",
+          reason: "Failed to get push notifications user data from storage",
+        };
+      }
+
+      return {
+        status: "error",
+        reason:
+          "Failed to get push notifications user data from storage and notification permission status",
+      };
+    }, [
+      initializeNotificationSettingsState,
+      notifications.areNotificationsAllowed,
+      setPermissionStatus,
+      syncOptOutState,
+      updatePushNotificationsDataOfUserInStateAndStore,
+    ]);
 
   const notifyFlowCompleted = useCallback(
     (source: NotificationsPromptAfterActionSource) => {
@@ -103,5 +179,6 @@ export function useNotificationsPromptTriggers() {
   return {
     notifyFlowCompleted,
     tryTriggerPushNotificationDrawerAfterInactivity,
+    initPushNotificationsData,
   };
 }
