+++ apps/ledger-live-mobile/src/mvvm/features/NotificationsPrompt/types.ts
@@ -15,12 +15,6 @@ export type DataOfUser = {
 
   // timestamp in ms of the last action user did (swap, receive, send, favorite, etc.)
   lastActionAt?: number;
-
-  // This old logic is helpful to know if the user has already opted out of notifications
-  /** If set, we will not prompt the push notification modal again before this date unless the user triggers it manually from the settings */
-  dateOfNextAllowedRequest?: Date;
-  /** Whether or not the user clicked on the "Maybe later" cta */
-  alreadyDelayedToLater?: boolean;
 };
 
 export type InitPushNotificationsDataResult =
